# web_nuxt/openapi.yaml
openapi: 3.0.0
info:
  title: StaffSec API.
  description: API документация для StaffSec.
  version: 1.0.0
servers:
  - url: http://{server}:{port}
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Используйте Bearer-токен в заголовке `Authorization`:  
        `Authorization: Bearer <token>`.  
        Токен получается через эндпоинт аутентификации.
...
paths:
  /route/anketa/self/{person_id}:
    get:
      summary: Переключение пользователя для редактирования анкеты
      security:
        - bearerAuth: []
      parameters:
        - name: person_id
          in: path
          required: true
          description: ID кандидата
          schema:
            type: integer
      responses:
        "201":
          description: Успешное переключение пользователя
          content:
            application/json:
              example: { "message": "success" }
        "500":
          description: Внутренняя ошибка
          content:
            application/json:
              example: { "message": "error" }
  /route/anketa/files/{person_id}:
    post:
      summary: Загрузка файлов на сервер
      security:
        - bearerAuth: []
      parameters:
        - name: person_id
          in: path
          required: true
          description: ID кандидата
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Успешная загрузка файлов
          content:
            application/json:
              example: { "message": "success" }
        "500":
          description: Внутренняя ошибка
          content:
            application/json:
              example: { "message": "error" }
  /route/anketa/json:
    post:
      summary: Создание или обновление анкеты из JSON
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: '../schemas/AnketaJson.json'
      responses:
        "201":
          description: Кандидата создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  person_id: {
                    type: integer
                  }
                  exists: {
                    type: boolean
                  }
...